+++ NEWS.md
@@ -24,6 +24,7 @@
 * Removed config option `maps_api_cdn_template`, reusing now instead `cdn_url`
 * Allow to create sync tables with a map if setting up onw from "connect dataset" from the Maps view
 * Added [#5975 Box integration](https://github.com/CartoDB/cartodb/issues/5975).
+* New rake to install in user or organization the geocoder extension
 
 
 3.11.0 (2015-09-09)